+++ docs/changelog.md
@@ -8,6 +8,9 @@
 - Play targetDestroyed sound when the computer reports another ship has been destroyed.
 - Add option to gag holographic avatar names in computer announcements (conf gag holo). This is disabled by default, so custom holographic avatar names will now be visible in computer announcements.
 
+### Changed
+- Wake on Red now also triggers on blue, purple, and orange alerts.
+
 ## Version 4.1.02
 
 ### New